Fix: even more code refactoring
